Updated the Twitch Rate Issue
Prior, every time any guild was updated, it used to reload the caches,
which included having the twitch listeners re-updated which resulted in
long wait times. Now, the twitch listeners are created on startup, and
only updated during the bot's online time. When the bot is restarted,
then it contacts the database and restarts the twitch listener.